src/main/java/me/jun/guestbook/service/PostService.java
package me.jun.guestbook.service;

import me.jun.guestbook.dao.PostRepository;
import me.jun.guestbook.domain.Post;
import me.jun.guestbook.dto.RequestPostId;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class PostService {

    private final PostRepository postRepository;

    @Autowired
    public PostService(PostRepository postRepository) {
        this.postRepository = postRepository;
    }

    public Post getPost(RequestPostId requestPostId) {
        Long id = requestPostId.getId();

        return postRepository.findById(id)
                .orElseThrow(IllegalAccessError::new);
    }
}
